Add files via upload
This project is called Lava Lamp. A lava lamp is a decorative lamp which consists of a bolus of a special colored wax mixture inside a glass vessel.
In total, the project consists of three parts: a plastic stand, glass vessel inside the plastic stand and a virtual arduino board as base light. When all pieces are combined it makes an aesthetically unique and beautiful lamp. 
The base light circuit consists of 8 red light bulbs which are turned on and off by a switch. The structure consists of 4 rings that are equally spaced and decrease in size as they descend and is supported by a rod which is also helpful in picking up the plastic structure. 
By: Sana Chaudhry, Maral Kenzhebek kyzy, Nursultan Ogonbasov, Adilet Kamilov.